Feat: Hive evolution and plotting

+++ src/scoutBee.h
@@ -26,17 +26,17 @@ class ScoutBee
 		int getChoice() const { return _choice; }
 		enum State { REST, SEARCH_NEW_NESTBOX, FIND_NESTBOX, BACK_TO_HOME, DANCE };
 		State getState() const { return _state; }
-		void setGene(float* gene);
+		void setGene(double* gene);
 
 		void draw();
 		__host__ __device__ void run(float random, float ratio, float hiveX, float hiveY, NestBox* nestBoxes, int qtyNestBoxes, float* choiceProb);
 	private:
 		// Gene
-		float* _gene;
-		float _randomChance;// Chance search new nestBox
-		float _followChance;// Chance follow other bee
-		float _linearDecay; // Linear supporting decay (0-1)
-		float _quadraticDecay; // Quadratic supporting decay (0-1)
+		double* _gene;
+		double _randomChance;// Chance search new nestBox
+		double _followChance;// Chance follow other bee
+		double _linearDecay; // Linear supporting decay (0-1)
+		double _quadraticDecay; // Quadratic supporting decay (0-1)
 
 		// Bee state
 		State _state;
